# Neptune.API/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Neptune.API",
    "version": "1.0"
  },
  "paths": {
...
    "/custom-attribute-types/{customAttributeTypeID}": {
      "get": {
        "tags": [
          "CustomAttributeType"
        ],
        "parameters": [
          {
            "name": "customAttributeTypeID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomAttributeTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomAttributeTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomAttributeTypeDto"
                }
              }
            }
          }
        }
      }
    },
    "/purpose/{customAttributeTypePurposeID}/treatmentBMPType/{treatmentBMPTypeID}": {
      "get": {
        "tags": [
          "CustomAttributeType"
        ],
        "parameters": [
          {
            "name": "customAttributeTypePurposeID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "treatmentBMPTypeID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomAttributeTypeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomAttributeTypeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomAttributeTypeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomAttributeTypeDto": {
        "type": "object",
        "properties": {
          "CustomAttributeTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "CustomAttributeTypeName": {
            "type": "string",
            "nullable": true
          },
          "CustomAttributeDataTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "MeasurementUnitTypeID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "IsRequired": {
            "type": "boolean"
          },
          "CustomAttributeTypeDescription": {
            "type": "string",
            "nullable": true
          },
          "CustomAttributeTypePurposeID": {
            "type": "integer",
            "format": "int32"
          },
          "CustomAttributeTypeOptionsSchema": {
            "type": "string",
            "nullable": true
          },
          "DataTypeDisplayName": {
            "type": "string",
            "nullable": true
          },
          "MeasurementUnitDisplayName": {
            "type": "string",
            "nullable": true
          },
          "Purpose": {
            "type": "string",
            "nullable": true
          },
          "CustomAttributeTypeSortOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}