# Neptune.Web/src/app/shared/generated/model/custom-attribute-type-dto.ts
/**
 * Neptune.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CustomAttributeTypeDto { 
    CustomAttributeTypeID?: number;
    CustomAttributeTypeName?: string;
    CustomAttributeDataTypeID?: number;
    MeasurementUnitTypeID?: number;
    IsRequired?: boolean;
    CustomAttributeTypeDescription?: string;
    CustomAttributeTypePurposeID?: number;
    CustomAttributeTypeOptionsSchema?: string;
    DataTypeDisplayName?: string;
    MeasurementUnitDisplayName?: string;
    Purpose?: string;
    CustomAttributeTypeSortOrder?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CustomAttributeTypeDtoForm { 
    CustomAttributeTypeID?: FormControl<number>;
    CustomAttributeTypeName?: FormControl<string>;
    CustomAttributeDataTypeID?: FormControl<number>;
    MeasurementUnitTypeID?: FormControl<number>;
    IsRequired?: FormControl<boolean>;
    CustomAttributeTypeDescription?: FormControl<string>;
    CustomAttributeTypePurposeID?: FormControl<number>;
    CustomAttributeTypeOptionsSchema?: FormControl<string>;
    DataTypeDisplayName?: FormControl<string>;
    MeasurementUnitDisplayName?: FormControl<string>;
    Purpose?: FormControl<string>;
    CustomAttributeTypeSortOrder?: FormControl<number>;
}

export class CustomAttributeTypeDtoFormControls { 
    public static CustomAttributeTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeDataTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MeasurementUnitTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsRequired = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypePurposeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeOptionsSchema = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static DataTypeDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MeasurementUnitDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Purpose = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomAttributeTypeSortOrder = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}